update to use BigDecimal throughout

+++ src/bad/robot/refactoring/chapter1/ChildrensPrice.java
@@ -4,8 +4,8 @@ import java.math.BigDecimal;
 
 public class ChildrensPrice extends Price {
     private static final int RENTAL_PERIOD = 3; // days
-    private static final double DAILY_LATE_FEE = 1.5;
-    private static final double BASE_CHARGE = 1.5;
+    private static final BigDecimal DAILY_LATE_FEE = BigDecimal.valueOf(1.5);
+    private static final BigDecimal BASE_CHARGE = BigDecimal.valueOf(1.5);
 
     @Override
     int getPriceCode() {
